feat(nsx-error): add error module
