+++ libs/nsx-error/src/lib/errors/entity-not-found-error/entity-not-found.constant.ts
@@ -1,16 +1,1 @@
-import { ID } from '@myancommerce/nox-common';
-import ApiError from '../../base.error';
-import {
-    ENTITY_NOT_FOUND_ERROR_CODE,
-    ENTITY_NOT_FOUND_ERROR_MESSAGE,
-} from './entity-not-found.error';
-
-export class EntityNotFoundError extends ApiError {
-    constructor(entityName: any, id: ID) {
-        super(
-            ENTITY_NOT_FOUND_ERROR_CODE,
-            { entityName, id },
-            ENTITY_NOT_FOUND_ERROR_MESSAGE,
-        );
-    }
-}
+export const ENTITY_NOT_FOUND_ERROR_CODE = 'error.entity-with-id-not-found';
