+++ libs/nsx-error/src/lib/errors/entity-not-found-error/entity-not-found.error.ts
@@ -1,2 +1,13 @@
-export const ENTITY_NOT_FOUND_ERROR_CODE = 'error.entity-with-id-not-found';
-export const ENTITY_NOT_FOUND_ERROR_MESSAGE = 'ENTITY_NOT_FOUND_ERROR_MESSAGE';
+import { ID } from '@myancommerce/nox-common';
+import ApiError from '../../base.error';
+import { ENTITY_NOT_FOUND_ERROR_CODE } from './entity-not-found.constant';
+
+export class EntityNotFoundError extends ApiError {
+    constructor(entityName: any, id: ID) {
+        super(
+            `Entity '${entityName}' with ID '${id}' not found.`,
+            { entityName, id },
+            ENTITY_NOT_FOUND_ERROR_CODE,
+        );
+    }
+}
